+++ client/src/components/ProtectedRoute.js
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { Route } from "react-router-dom";
 import Login from "components/Login";
-import UserContext from "../contexts/UserContext";
+import UserContext from "contexts/UserContext";
 
 const ProtectedRoute = ({ component: Component, ...rest }) => {
   const value = useContext(UserContext);
