+++ client/src/pages/Dashboard.js
@@ -1,24 +1,16 @@
 import React, { useState } from "react";
-import { Box } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import Header from "components/Header";
-import ShoppingLists from "pages/ShoppingLists";
-import Friends from "pages/Friends";
+
+import Header from "elements/Header";
+import Page from "elements/Page";
+import Body from "elements/Body";
+import HeaderTabs from "components/HeaderTabs";
+
+import ShoppingLists from "body/ShoppingLists";
+import Friends from "body/Friends";
 
 const useStyles = makeStyles((theme) => ({
-  dashContainer: {
-    position: "absolute",
-    top: "0",
-    left: "0",
-    right: "0",
-    minHeight: "100vh",
-    display: "flex",
-    flexDirection: "column",
-  },
-  bodyContainer: {
-    flex: "1",
-    backgroundColor: "#f8f8f8",
-  },
+  //
 }));
 
 const Dashboard = () => {
@@ -27,20 +19,22 @@ const Dashboard = () => {
   const classes = useStyles();
 
   return (
-    <Box className={classes.dashContainer}>
-      <Header
-        {...{
-          selectedPage,
-          setSelectedPage,
-          notificationsOpen,
-          setNotificationsOpen,
-        }}
-      />
-      <Box className={classes.bodyContainer}>
+    <Page>
+      <Header>
+        <HeaderTabs
+          {...{
+            selectedPage,
+            setSelectedPage,
+            notificationsOpen,
+            setNotificationsOpen,
+          }}
+        />
+      </Header>
+      <Body>
         {selectedPage === 0 && <ShoppingLists />}
         {selectedPage === 1 && <Friends />}
-      </Box>
-    </Box>
+      </Body>
+    </Page>
   );
 };
